[build] Finished the Python-based build system
This commit introduces further changes in `y.py`.
From now on, it is the primary build system for Esque.
While some may wonder why this is neccessary, a simple
look at more complex projects, such as the linux kernel
serves an answer quickly. The Makefiles have devolved,
from a once great build system, to an unreadable
mess. Python is in today's age, in my opinion, the
Lingua Franca of programmers. I have yet to meet
a person who did not understand, at least partly, python
code.
A kernel on it's own is a complex project, one that
we should not try to overcomplicate by using a build-
system that is guaranteed to be unreadable for many
after a certain amount of time.