[feat] Started working on Messages for the IPC
The IPC is going to be a vital part of Esque. In it,
communications happen via so-called messages.
The following describes a message:
Message {
	source: Process,
	destination: Process,
	typ: i32,
	content: MessageContent
}
where `MessageContent` is a union with *all* possible message
types in it.

This system offers one big pro:
It is not too memory-consuming while still offering
flexibility.

Cons:
We must *trust* the user to know which field of the union
to use for which call. But, in real life, we trust
the user too, when they make a system call.