[test] Added a full testing framework for the kernel
This new commit adds a whole new testing framework to the kernel.
From now on, commits will be tested. A successful test may look like this:

```
\#[esqtest::test]
pub fn check_macros() {

    check_eq!(0, 0); // Equal - OK
    check!(true); // Any boolean expression - OK
    check_neq!(1, 2); // Not equal - OK
    check_gt!(1, 0); // Greater Than - OK
    check_lt!(0, 1); // Less than - OK
    check_ge!(1, 1); // Greater or equal - OK
    check_le!(100, 1000); // Less or equal - OK
    all_good!(); // Simply returns 0
}
```

The following is a test that will always fail:
```
\#[esqtest::test]
pub fn i_always_fail() {
    check!(false); // Returns one from function
    all_good!();
}
```

All of those tests return numbers. If the number is 0 the test
succeeded, if it is not, the test failed.

If the `harsh-tests` feature is enabled, QEMU will exit if any error
occured.
It will exit with

```
- 0x42 if *all* tests were successful
- 0x43 if *more than* 50% of the tests failed
- 0x44 if *less than* 50%, but *more than* 0% of the tests failed

This framework has been integrated into the CI